+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+## 0.42.0
+
+### New Features ✨
+
+- (billing) Protos for platform Pay Now webhook loop by @armcknight in [#345](https://github.com/getsentry/sentry-protos/pull/345)
+
+## 0.41.2
+
+### Internal Changes 🔧
+
+- Add total quantity by @noahsmartin in [#349](https://github.com/getsentry/sentry-protos/pull/349)
+
+## 0.41.1
+
+### Internal Changes 🔧
+
+- (billing-platform) Get all contracts for organization by @brendanhsentry in [#348](https://github.com/getsentry/sentry-protos/pull/348)
+
 ## 0.41.0
 
 ### New Features ✨